+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/DataClassExtensions.kt
@@ -56,3 +56,28 @@ inline fun <reified T> toStringByRef(target: T?, delimiter: String = ", ", prefi
 		.let { if(omitNulls) it.filterValuesNotNull() else it }
 		.joinToString(delimiter, className + prefix, postfix) { (k, v) -> "$k=${v.smartToString(deepOperation)}" }
 }
+
+
+/**智能地判断两个对象是否相等。特殊对待数组类型，默认递归执行操作。*/
+@PublishedApi
+internal fun Any?.smartEquals(other: Any?, deepOperation: Boolean = true) = when {
+	this !is Array<*> || other !is Array<*> -> this == other
+	!deepOperation -> this contentEquals other
+	else -> this contentDeepEquals other
+}
+
+/**智能地得到当前对象的哈希码。特殊对待数组类型，默认递归执行操作。*/
+@PublishedApi
+internal fun Any?.smartHashCode(deepOperation: Boolean = true) = when {
+	this !is Array<*> -> this.hashCode()
+	!deepOperation -> this.contentHashCode()
+	else -> this.contentDeepHashCode()
+}
+
+/**智能地将当前对象转化为字符串。特殊对待数组类型，默认递归执行操作。*/
+@PublishedApi
+internal fun Any?.smartToString(deepOperation: Boolean = true) = when {
+	this !is Array<*> -> this.toString()
+	!deepOperation -> this.contentToString()
+	else -> this.contentDeepToString()
+}